subfolders + tmpdir name (sha2 content)

+++ translator/test_infra/src/sol.rs
@@ -1,9 +1,10 @@
-use anyhow::{ensure, Error};
 use std::io::Write;
 use std::process::Command;
 use std::sync::Arc;
 use std::{fs, io};
-use tempfile::TempDir;
+
+use anyhow::{ensure, Error};
+use sha3::{Digest, Sha3_256};
 
 #[derive(Debug, Clone)]
 pub struct Evm {
@@ -27,29 +28,30 @@ impl Evm {
 }
 
 pub fn build_sol(sol: &[u8]) -> Result<Evm, Error> {
-    let tmp_dir = TempDir::new()?;
-    let tmp_dir = tmp_dir.path().join(format!("{}", rand::random::<u64>()));
-    fs::create_dir_all(&tmp_dir)?;
+    let tmp_dir = std::env::temp_dir().join(sha_name(sol));
+    if !tmp_dir.exists() {
+        fs::create_dir_all(&tmp_dir)?;
 
-    let contract = tmp_dir.join("contract.sol");
-    fs::write(&contract, sol)?;
+        let contract = tmp_dir.join("contract.sol");
+        fs::write(&contract, sol)?;
 
-    let output = Command::new("solc")
-        .current_dir(tmp_dir.as_path())
-        .arg("-o")
-        .arg(tmp_dir.as_path())
-        .arg("--bin")
-        .arg("--abi")
-        .arg(contract.as_path())
-        .output()?;
-    io::stdout().write_all(&output.stdout)?;
-    io::stderr().write_all(&output.stderr)?;
-    ensure!(
-        output.status.success(),
-        "Compilation error:\n{}",
-        String::from_utf8(output.stderr).unwrap_or_default()
-    );
-    fs::remove_file(&contract)?;
+        let output = Command::new("solc")
+            .current_dir(tmp_dir.as_path())
+            .arg("-o")
+            .arg(tmp_dir.as_path())
+            .arg("--bin")
+            .arg("--abi")
+            .arg(contract.as_path())
+            .output()?;
+        io::stdout().write_all(&output.stdout)?;
+        io::stderr().write_all(&output.stderr)?;
+        ensure!(
+            output.status.success(),
+            "Compilation error:\n{}",
+            String::from_utf8(output.stderr).unwrap_or_default()
+        );
+        fs::remove_file(&contract)?;
+    }
 
     let dir = fs::read_dir(tmp_dir)?;
     let dir = dir.into_iter().collect::<Result<Vec<_>, _>>()?;
@@ -84,4 +86,12 @@ pub fn build_sol(sol: &[u8]) -> Result<Evm, Error> {
         abi: Arc::new(abi),
         bin: Arc::new(bin),
     })
+}
+
+fn sha_name(cont: &[u8]) -> String {
+    let mut hasher = Sha3_256::new();
+    hasher.update(cont);
+    let value_hash = hasher.finalize();
+    let hash: [u8; 32] = value_hash.as_slice().try_into().expect("Wrong length");
+    format!(".{}", base32::encode(base32::Alphabet::Crockford, &hash))
 }
